minor check added to make sure that elements_per_page for bitmap page fit on 32 bit integer

+++ inc/tupleindexer/bitmap_page/bitmap_page.h
@@ -10,7 +10,8 @@
 // this tuple has a tuple_def having a fixed length array of non-nullable bit-fields that has bits_per_field in range [1, 64], number of bits
 // the number of elements on the page depends on the size of the page and is also returned
 // any access to bit_fields (using the get and set functions below) outside of (*elements_per_page) is an an out-of-bounds access
-tuple_def* get_tuple_definition_for_bitmap_page(const page_access_specs* pas_p, uint8_t bits_per_field, uint64_t* elements_per_page);
+// note: this function may also return NULL, if the elements_per_page overflows a 32 bit unsigned integer
+tuple_def* get_tuple_definition_for_bitmap_page(const page_access_specs* pas_p, uint8_t bits_per_field, uint32_t* elements_per_page);
 
 persistent_page get_new_bitmap_page_with_write_lock(const page_access_specs* pas_p, const tuple_def* tpl_d, const page_access_methods* pam_p, const page_modification_methods* pmm_p, const void* transaction_id, int* abort_error);
